refactored test input files to be stored in a subpackage, renamed test folders, deleted redundant casting
