--- src/Test/OptionsPackageTest/PostTest.java
@@ -1,4 +1,0 @@
-package OptionsPackageTest;
-
-public class PostTest {
-}
