+++ src/Test/loginFunctionsTest/SignInGatewayTest.java
@@ -1,30 +1,30 @@
-package UserFunctionsTest;
+package loginFunctionsTest;
 
 import settingsFunctions.DeleteUserGateway;
 import userFunctions.User;
 import loginFunctions.GetUserGateway;
-import userFunctions.SaveUserChangesGateways;
+import loginFunctions.SignUpGateway;
 import org.junit.Test;
 
 import java.io.IOException;
 
 import static org.junit.Assert.assertEquals;
 
-public class SaveUserChangesGatewaysTest {
-    SaveUserChangesGateways saveUserChangesGateways = new SaveUserChangesGateways();
+public class SignInGatewayTest {
+
     GetUserGateway getUserGateway = new GetUserGateway();
+    SignUpGateway signUpGateway = new SignUpGateway();
     DeleteUserGateway deleteUserGateway = new DeleteUserGateway();
 
     @Test
-    public void setSaveUserChangesGatewaysTest() throws IOException, ClassNotFoundException {
+    public void getUserBasicTest() throws IOException, ClassNotFoundException {
         User newUser = new User("albert");
-        String username = newUser.getUsername();
-        saveUserChangesGateways.save(username, newUser);
+        signUpGateway.allowSignUp(newUser.getUsername(), newUser);
 
         User expectedUser = getUserGateway.getUser("albert");
         assertEquals(newUser.getUsername(), expectedUser.getUsername());
         deleteUserGateway.deleteUser("albert");
 
     }
-}
 
+}
